+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Pirate's treasure".
 - Tests for "Simple awalé".
 - Tests for "Simple load balancing".
+- Tests for "Are the clumps normal".
 
 ## [3.8.0] - 2022-04-30
 ### Added